Document Picker and Document Preview (#13)
* Document Picker

added a view modifier for selecting a file (pdf, image) on the device and expose its url to the app

Document Viewer

added a view modifier for previewing a selected file (pdf, image) that's availble in the app. It could be a local downladed file or a file selected with the document picker

ReadMe

updated the readme to describe the two new view modifiers

* Document Preview

Fixed a variable name after a refactoring

* DocumetPicker

simplified the implementation of cache copyiing using the BufferedInputStream copyTo method

added skip nobridge

* Document Picker

Added output and input close at the end of the file copy as stated in : https://kotlinlang.org/api/core/kotlin-stdlib/kotlin.io/copy-to.html

* MediaPicker

fixed a bug that when using for the first time after showing the permission request dialog on Andoroid users wont be able to tap on the button to activate the picker unless they reload the view, since the `isPresented` flag was still set to true.

* UTType stub

Created a UTType stub for Android for letting user choose what kind of files select with the document picker.

Note: it compiles fine, but when using the view extension `.withDocumentPicker` in an actual project it gives a compilation error `Argument type mismatch: actual type is 'skip.lib.Array<Element>', but 'skip.lib.Array<skip.kit.UTType>'`

* Update UTType and docs cleanup

* Add !SKIP_BRIDGE header check

* Do not bridge unsupported optional UTType init

---------

Co-authored-by: Marc Prud'hommeaux <mwp1@cornell.edu>
Co-authored-by: Marc Prud'hommeaux <marc@skip.tools>